% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{add_group_stats}
\alias{add_group_stats}
\title{Add group-wise cumulative statistics}
\usage{
add_group_stats(dt, group_var, value_var, fun = cumsum, prefix = "c")
}
\arguments{
\item{dt}{A data.table (or convert beforehand with \code{data.table::setDT()}).}

\item{group_var}{Columns to group by.}

\item{value_var}{Columns to apply the function to.}

\item{fun}{A function to apply. Default is \code{cumsum}}

\item{prefix}{A new column's perfix string (if the value_var is "loss",
the new column is "closs")}
}
\value{
No return values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Adds statistic columns computed \strong{within groups}, typically using
cumulative functions such as \code{\link[=cumsum]{cumsum()}}, \code{\link[=cumprod]{cumprod()}}, \code{\link[=cummax]{cummax()}},
or \code{\link[=cummin]{cummin()}}.
}
\examples{
\donttest{
# Cumulative sum within groups
dt <- data.table::as.data.table(mtcars)
add_group_stats(dt, group_var = .(cyl), value_var = c(hp, mpg), fun = cumsum)
}

}
