% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{add_mon}
\alias{add_mon}
\title{Add months}
\usage{
add_mon(date, mon)
}
\arguments{
\item{date}{A date}

\item{mon}{A number of months to be added}
}
\value{
A date
}
\description{
Add months to the date.
}
\examples{
# add months
\donttest{
date <- Sys.Date()
add_mon(date, 3)}

}
