% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{add_year}
\alias{add_year}
\title{Add years to a date}
\usage{
add_year(date, year)
}
\arguments{
\item{date}{A date-like object (coercible via \code{\link[=as.POSIXlt]{as.POSIXlt()}}).}

\item{year}{Integer number of years to add (can be negative).}
}
\value{
A Date with the years added.
}
\description{
Add a specified number of years to a date-like object. Negative values
subtract years. Uses base rollover via POSIXlt.
}
\examples{
\donttest{
# Add years
date <- Sys.Date()
add_year(date, 3)
}

}
