% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_comma}
\alias{as_comma}
\title{As comma applied label}
\usage{
as_comma(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a string vector
}
\description{
Convert a numeric vector to a comma applied string vector.
}
\examples{
# convert to a comma applied string vector
\donttest{as_comma(c(123456, 234567))}


}
