% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_numeric}
\alias{as_numeric}
\alias{as_logical}
\alias{as_integer}
\alias{as_double}
\alias{as_character}
\title{As functions}
\usage{
as_numeric(x)

as_logical(x)

as_integer(x)

as_double(x)

as_character(x)
}
\arguments{
\item{x}{a vector or matrix}
}
\value{
a converted vector or matrix
}
\description{
As logical, integer, numeric, character maintaining the original shape.
}
