% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_class}
\alias{assert_class}
\title{Assert class}
\usage{
assert_class(obj, class)
}
\arguments{
\item{obj}{object}

\item{class}{object class}
}
\value{
no return
}
\description{
Assert object class.
}
\examples{
# assert object class
\donttest{assert_class(cars, "data.frame")}

}
