% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_length}
\alias{assert_length}
\title{Assert nonzero length}
\usage{
assert_length(x)
}
\arguments{
\item{x}{An object of class \code{logical}, \code{integer}, \code{numeric}, \code{character},
\code{complex}, \code{raw}, \code{Date}, \code{POSIXt}, \code{factor}, or \code{ordered}.}
}
\value{
Invisibly returns \code{TRUE} on success; otherwise throws an error.
}
\description{
Verifies that an object has nonzero length. On failure, throws an error.
}
\examples{
\dontrun{
assert_length(1:3)
assert_length(numeric()) # errors
}

}
\seealso{
\code{\link[rlang:has_length]{rlang::has_length()}}
}
