% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_rows}
\alias{assert_rows}
\title{Assert that a data frame has rows}
\usage{
assert_rows(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
Invisibly returns \code{TRUE} on success; otherwise throws an error.
}
\description{
Verifies that a data frame has at least one row. On failure, throws an error.
}
\examples{
\dontrun{
assert_rows(data.frame(1))
assert_rows(data.frame()) # error
}

}
