% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_type}
\alias{assert_type}
\title{assert type}
\usage{
assert_type(obj, type)
}
\arguments{
\item{obj}{object}

\item{type}{A type string}
}
\value{
no return
}
\description{
Check the type of object.
}
