% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{bmonth}
\alias{bmonth}
\alias{emonth}
\title{Beginning of the month, End of the month}
\usage{
bmonth(date)

emonth(date)
}
\arguments{
\item{date}{A date}
}
\value{
A date
}
\description{
Get the beginning of the month.
}
\examples{
# the beginning of the month
\donttest{
bmonth(Sys.Date())}

# the end of the month
\donttest{
emonth(Sys.Date())}

}
