% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{col_max}
\alias{col_max}
\alias{col_min}
\alias{col_sum}
\title{max, min, sum on each column of a matrix}
\usage{
col_max(x)

col_min(x)

col_sum(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A vector of min, max, sum of values on each column of a matrix
}
\description{
Get max, min, sum of values on each column of a matrix.
}
\examples{
# column max
\donttest{x <- matrix(c(1:9), ncol = 3)
col_max(x)}

# column min
\donttest{x <- matrix(c(1:9), ncol = 3)
col_min(x)}

# column sum
\donttest{x <- matrix(c(1:9), ncol = 3)
col_sum(x)}

}
