% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{count_pattern}
\alias{count_pattern}
\title{Count pattern matched strings}
\usage{
count_pattern(pattern, x, ignore.case = FALSE)
}
\arguments{
\item{pattern}{a string containing a \link{regular expression}}

\item{x}{a string}

\item{ignore.case}{if \code{FALSE}, the pattern matching is case sensitive and if \code{TRUE}, case is ignored during matching.}
}
\value{
a string vector
}
\description{
Count pattern matched strings from a string vector.
}
\examples{
# count pattern matched strings from a string vector
\donttest{count_pattern(pattern = "c", c("a|b|c", "a|c|c"))}

}
