% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{create_library_r}
\alias{create_library_r}
\title{Create library.R}
\usage{
create_library_r(path = "library.R")
}
\arguments{
\item{path}{a string value specifying the path of the file.}
}
\value{
no return value
}
\description{
Create library.R file
}
\examples{
# create library.R
\donttest{create_library_r("library.R")}

}
