% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{data_xlsx}
\alias{data_xlsx}
\title{Deprecated: data_xlsx()}
\usage{
data_xlsx(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=write_data_xlsx]{write_data_xlsx()}}.}
}
\value{
Same return value as \code{\link[=write_data_xlsx]{write_data_xlsx()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=write_data_xlsx]{write_data_xlsx()}} instead.
}
\seealso{
\code{\link[=write_data_xlsx]{write_data_xlsx()}}
}
