% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{del_pattern}
\alias{del_pattern}
\title{Delete patterns}
\usage{
del_pattern(pattern, x)
}
\arguments{
\item{pattern}{a string containing a \link{regular expression}}

\item{x}{a string vector}
}
\value{
a string vector
}
\description{
Delete patterns from a string vector.
}
\examples{
# delete patterns from a string vector
\donttest{del_pattern(pattern = "c", c("abc", "acc"))}

}
