% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{desub}
\alias{desub}
\title{Desub}
\usage{
desub(x)
}
\arguments{
\item{x}{an expression that can be a string vector}
}
\value{
a string vector
}
\description{
This function operates like \code{deparse(substitute(x))} inside the functions.
}
\examples{
# desub
\donttest{f1 <- function(a) desub(a)
f2 <- function(b) f1(b)
f3 <- function(c) f2(c)
f4 <- function(d) f3(d)
f5 <- function(e) f4(e)
desub(iris) # iris
f1(iris) # iris
f2(iris) # iris
f3(iris) # iris
f4(iris) # iris
f5(iris) # iris}

}
