% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{devars}
\alias{devars}
\title{devars}
\usage{
devars(x)
}
\arguments{
\item{x}{A string, vector or list expression that can be a string vector}
}
\value{
A string vector
}
\description{
This function operates like \code{deparse(substitute(x))} inside the functions.
}
\examples{
# deparse(substitute(x))
\donttest{devars(expression)
devars(c(expression, string))
devars(list(expression, string))
devars(.(expression, string))}

}
