% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_cols}
\alias{diff_cols}
\title{Different columns}
\usage{
diff_cols(df, cols)
}
\arguments{
\item{df}{a data.frame}

\item{cols}{a string vector specifying columns}
}
\value{
a string vector
}
\description{
Columns that the data frame does not contain
}
\examples{
# different columns
\donttest{diff_cols(mtcars, c("mpg", "cyl", "disp", "hp", "drat"))}


}
