% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{dolock}
\alias{dolock}
\alias{unlock}
\title{dolock, unlock}
\usage{
dolock(obj, key)

unlock(obj, key)
}
\arguments{
\item{obj}{An object to save}

\item{key}{A string specifying password}
}
\description{
lock and unlock data files for security.
}
\examples{
## dolock
\dontrun{dolock("your data", "your password")}

## unlock
\dontrun{unlock("your data", "your password")}

}
