% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{draw_xlsx}
\alias{draw_xlsx}
\title{Draw images in an excel file}
\usage{
draw_xlsx(
  image,
  file,
  rc = c(1L, 1L),
  width = 12,
  height = 6,
  overwrite = FALSE
)
}
\arguments{
\item{image}{image file}

\item{file}{A file path}

\item{rc}{A vector of starting point row and column, default c(1L, 1L)}

\item{width}{image width}

\item{height}{image height}

\item{overwrite}{A boolean value that overwrite or not}
}
\value{
no return
}
\description{
Draw images in an excel file.
}
\examples{
# draw images in an excel file
\dontrun{
draw_xlsx(list(image1, image2), "image.xlsx")}

}
