% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{fill_zero_not_first_pos}
\alias{fill_zero_not_first_pos}
\alias{set_zero_not_first_pos}
\title{fill with or set to zero except for the first positive values}
\usage{
fill_zero_not_first_pos(x, cols)

set_zero_not_first_pos(x, cols)
}
\arguments{
\item{x}{A numerical matrix}

\item{cols}{A boolean vector. for example, if you want select first and third
columns for 3 X 3 matrix, c(1, 0, 1).}
}
\value{
A matrix with only first positive value for each row names, and the
other values become 0.
}
\description{
Fill with or set to zero except for the first positive values of a numerical
matrix by rownames
}
\examples{
# fill with zero except for the first positive values of a numerical matrix
# by rownames
\donttest{x <- matrix(c(-1, 2, 0, -2, -1, 5, 2, -2, 3), nrow = 3)
set_rownames(x, c(1, 1, 2))
fill_zero_not_first_pos(x)}

# set to zero except for the first positive values of a numerical matrix by
# rownames
\donttest{x <- matrix(c(-1, 2, 0, -2, -1, 5, 2, -2, 3), nrow = 3)
set_rownames(x, c(1, 1, 2))
set_zero_not_first_pos(x)}

}
