% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{freq_prop}
\alias{freq_prop}
\title{Compute frequency proportions}
\usage{
freq_prop(x)
}
\arguments{
\item{x}{A vector (numeric, character, or factor).}
}
\value{
A data.table with unique values and their proportions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Given a vector, compute the relative frequency of each unique value
and return the result as a data.frame.
}
\examples{
\donttest{
# Frequency proportions
x <- sample(1:10, 1000, replace = TRUE)
freq_prop(x)
}

}
