% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_copied_dt}
\alias{get_copied_dt}
\title{Get a copied data.table}
\usage{
get_copied_dt(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a copied data.table
}
\description{
Get a copied data.table.
}
\examples{
# get copied data.table
\donttest{df <- data.frame(x = 1:3, y = c("a", "b", "c"))
get_copied_dt(df)}

}
