% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_pos.R
\name{get_first_pos}
\alias{get_first_pos}
\alias{set_first_pos}
\title{get or set first positive values}
\usage{
get_first_pos(x, cols)

set_first_pos(x, cols)
}
\arguments{
\item{x}{A numerical matrix}

\item{cols}{A boolean vector. for example, if you want select first and third
columns for 3 X 3 matrix, c(1, 0, 1).}
}
\value{
A matrix with only first positive value for each row names, and the
other values become 0.
}
\description{
Get or set first positive values of a numerical matrix by rownames
}
\examples{
# get first positive values of a numerical matrix by rownames
\donttest{x <- matrix(c(-1, 2, 0, -2, -1, 5, 2, -2, 3), nrow = 3)
set_rownames(x, c(1, 1, 2))
get_first_pos(x)}

# set first positive values of a numerical matrix by rownames
\donttest{x <- matrix(c(-1, 2, 0, -2, -1, 5, 2, -2, 3), nrow = 3)
set_rownames(x, c(1, 1, 2))
set_first_pos(x)}

}
