% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{get_japanese_cols}
\alias{get_japanese_cols}
\title{Get japanese columns?}
\usage{
get_japanese_cols(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
japanese column names
}
\description{
Get japanese columns?
}
