% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{get_pattern}
\alias{get_pattern}
\title{Get first match}
\usage{
get_pattern(pattern, x, ignore.case = TRUE)
}
\arguments{
\item{pattern}{A string containing a regular expression.}

\item{x}{A character vector.}

\item{ignore.case}{Logical; if \code{TRUE}, case is ignored. Default: \code{TRUE}.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Extract the first regex match from each string (empty string if no match).
}
\examples{
\donttest{get_pattern(pattern = "c", c("a|b|c", "a|c|c"))}

}
