% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{get_prop}
\alias{get_prop}
\title{Get proportion from a vector}
\usage{
get_prop(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a proportion data.frame
}
\description{
Get proportion from a vector.
}
\examples{
\dontrun{
get_prop(sample(1:10, 1000, replace = TRUE))}

}
