% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{get_ptr}
\alias{get_ptr}
\title{Get external pointer}
\usage{
get_ptr(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
No return value.
}
\description{
Get external pointer
}
\examples{
# get pointer
\dontrun{
df <- data.frame(x = c(1:3), y = c("a", "b", "c"))
get_ptr(df)
df <- setalloccol(df)
get_ptr(df)}

}
