% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{get_stat_by}
\alias{get_stat_by}
\title{Get statistic by groups}
\usage{
get_stat_by(df, group_var, value_var, fun = sum)
}
\arguments{
\item{df}{a data.frame}

\item{group_var}{names of columns for grouping}

\item{value_var}{names of columns to be applied for statistics.}

\item{fun}{a function applying to columns}
}
\value{
a grouped data.frame
}
\description{
Get statistic column by groups
}
\examples{
# get statistics by groups
\dontrun{
set_ptr(mtcars)
get_stat_by(mtcars, .(cyl, vs), value_var = .(hp, drat), fun = sum)}

}
