% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{grepl_and}
\alias{grepl_and}
\title{grepl with \code{And} conditions}
\usage{
grepl_and(pattern, x, ignore.case = TRUE)
}
\arguments{
\item{pattern}{a character vector containing regular expressions}

\item{x}{a character vector}

\item{ignore.case}{if \code{FALSE}, the pattern matching is case sensitive and
if \code{TRUE}, case is ignored during matching.}
}
\value{
a boolean vector
}
\description{
grepl with \code{And} conditions.
}
\examples{
# grepl with "And" conditions
\donttest{string <- "abcde12345"
grepl_and(c("a", "c", "e"), string)
grepl_and(c("a", "c", "[1-5]"), string)
grepl_and(c("a", "c", "z"), string)}

}
