% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{has_attrs}
\alias{has_attrs}
\title{Check if an object has specific attributes}
\usage{
has_attrs(obj, attrs)
}
\arguments{
\item{obj}{An R object.}

\item{attrs}{A character vector of attribute names to check.}
}
\value{
Logical scalar: \code{TRUE} if all attributes are present, otherwise \code{FALSE}.
}
\description{
Returns \code{TRUE} if an object has \strong{all} of the specified attributes,
otherwise \code{FALSE}. Unlike \code{\link[=assert_attrs]{assert_attrs()}}, this never throws an error.
}
\examples{
has_attrs(mtcars, c("names", "class"))    # TRUE
has_attrs(mtcars, c("foo", "bar"))        # FALSE

}
\seealso{
\code{\link[=assert_attrs]{assert_attrs()}}
}
