% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_cols}
\alias{has_cols}
\title{Check if a data frame has specific columns}
\usage{
has_cols(df, cols, error_raise = FALSE)
}
\arguments{
\item{df}{A \code{data.frame}.}

\item{cols}{A character vector of column names to check.}

\item{error_raise}{Logical; if \code{TRUE}, raise an error when columns are missing.}
}
\value{
A logical scalar (\code{TRUE} if all specified columns exist, otherwise \code{FALSE}).
}
\description{
Test whether a data frame contains all of the specified columns.
Can return a logical value or raise an error if \code{error_raise = TRUE}.
}
\examples{
\dontrun{
# Has columns
has_cols(mtcars, c("cyl", "disp"))

# Raise an error
has_cols(mtcars, c("cyl", "iris"), error_raise = TRUE)
}

}
