% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_len}
\alias{has_len}
\title{Check if an object has nonzero length}
\usage{
has_len(x, error_raise = FALSE)
}
\arguments{
\item{x}{An object of class \code{character}, \code{integer}, \code{numeric}, \code{Date}, or \code{POSIXt}.}

\item{error_raise}{Logical; if \code{TRUE}, raise an error if the object has zero length.}
}
\value{
A logical scalar (\code{TRUE} if \code{x} has nonzero length, otherwise \code{FALSE}).
}
\description{
Test whether an object has nonzero length. Can return a logical value
or raise an error if \code{error_raise = TRUE}.
}
\examples{
\dontrun{
# Has a length
has_len(c(numeric(), character()))

# Raise an error
has_len(c(numeric(), character()), error_raise = TRUE)
}

}
\seealso{
\code{\link[rlang:has_length]{rlang::has_length()}}
}
