% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_rows}
\alias{has_rows}
\title{Check if a data frame has rows}
\usage{
has_rows(df, error_raise = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{error_raise}{Logical; if \code{TRUE}, raise an error if the data frame has no rows.}
}
\value{
A logical scalar (\code{TRUE} if the data frame has rows, otherwise \code{FALSE}).
}
\description{
Test whether a data frame has at least one row. Can return a logical value
or raise an error if \code{error_raise = TRUE}.
}
\examples{
\dontrun{
df <- data.frame()

# Has columns
has_rows(df) # FALSE

# Raises an error
has_rows(df, error_raise = TRUE)
}

}
