% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icol}
\alias{icol}
\title{Get column indices}
\usage{
icol(x, cols)
}
\arguments{
\item{x}{A data.frame.}

\item{cols}{A character vector of column names.}
}
\value{
An integer vector of column indices.
}
\description{
Return the column numbers corresponding to given column names.
}
\examples{
\donttest{
# Column numbers for selected names
icol(mtcars, c("disp", "drat", "qsec", "am", "carb"))
}

}
