% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{image_xlsx}
\alias{image_xlsx}
\title{Insert images in an excel file}
\usage{
image_xlsx(
  images,
  file,
  rc = c(1L, 1L),
  width = 12,
  height = 6,
  dpi = 300,
  overwrite = FALSE
)
}
\arguments{
\item{images}{image file paths}

\item{file}{a file path}

\item{rc}{a vector of starting point row and column, default c(1L, 1L)}

\item{width}{image width}

\item{height}{image height}

\item{dpi}{image resolution}

\item{overwrite}{A boolean value that overwrite or not}
}
\value{
no return
}
\description{
Insert images in an excel file.
}
\examples{
# insert images in an excel file
\dontrun{
image_xlsx(list(image1, image2), "image.xlsx")}

}
