% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{interleave}
\alias{interleave}
\title{Interleave two vectors}
\usage{
interleave(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A vector containing the elements of \code{x} and \code{y} interleaved.
The type will follow the usual R coercion rules when combining vectors.
}
\description{
Combine two vectors by interleaving their elements. That is, take one element
from \code{x}, then one from \code{y}, alternating until both vectors are exhausted.
}
\examples{
\donttest{
# Interleave two numeric vectors
x <- c(1, 3, 5, 7)
y <- c(2, 4, 6, 8)
interleave(x, y)
}

}
