% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.null.externalptr}
\alias{is.null.externalptr}
\title{Is a null external pointer?}
\usage{
is.null.externalptr(pointer)
}
\arguments{
\item{pointer}{an externalptr object}
}
\value{
a logical value whether it is null external pointer or not.
}
\description{
Is a null external pointer?
}
\examples{
# is null external pointer?
\dontrun{
p <- new("externalptr")
is.null.externalptr(p)}

}
