% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{is_date_format}
\alias{is_date_format}
\title{Is a date format?}
\usage{
is_date_format(x)
}
\arguments{
\item{x}{value to check}
}
\value{
a logical whether it's a date format or not.
}
\description{
Is a date or date format vector?
}
