% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{is_null_externalptr}
\alias{is_null_externalptr}
\title{Is a NULL external pointer?}
\usage{
is_null_externalptr(ptr)
}
\arguments{
\item{ptr}{An externalptr object.}
}
\value{
A single logical value: \code{TRUE} if \code{ptr} is a NULL external pointer,
otherwise \code{FALSE}.
}
\description{
Checks whether an object of type \code{externalptr} is a \strong{NULL} external pointer.
This does \strong{not} test for R's \code{NULL}, but for a C-level null pointer behind
an \code{externalptr}.
}
\examples{
# is null external pointer?
\dontrun{
p <- new("externalptr")
is_null_externalptr(p)
}

}
\seealso{
\code{\link[=get_ptr]{get_ptr()}}, \code{\link[=drop_ptr]{drop_ptr()}}
}
