% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_attr}
\alias{match_attr}
\title{Find matching attributes}
\usage{
match_attr(x, name)
}
\arguments{
\item{x}{Any R object (e.g., \code{list}, \code{data.frame}, \code{data.table}).}

\item{name}{Character vector of attribute names to match.}
}
\value{
A character vector of matching attribute names.
}
\description{
Returns the names of attributes in \code{x} that match the specified names.
}
\examples{
\donttest{
# Find attributes by name
match_attr(iris, c("class", "names"))
}

}
