% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_cols}
\alias{match_cols}
\title{Match columns}
\usage{
match_cols(df, cols)
}
\arguments{
\item{df}{a data frame}

\item{cols}{a string vector specifying columns}
}
\value{
a string vector
}
\description{
Get matched columns from a data frame.
}
\examples{
# match columns
\donttest{df <- data.frame(x = c(1, 2, 3), y = c("A", "B", "C"), z = c(4, 5, 6))
match_cols(df, c("x", "z"))}

}
