% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matXcol}
\alias{matXcol}
\title{Matrix X Column vector}
\usage{
matXcol(mat, vec)
}
\arguments{
\item{mat}{a numeric matrix}

\item{vec}{a numeric vector}
}
\value{
no return value
}
\description{
Multiply a matrix and a column vector and allocate the result to the matrix memory.
}
