% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matXnum}
\alias{matXnum}
\title{Matrix X Number}
\usage{
matXnum(mat, num)
}
\arguments{
\item{mat}{a numeric matrix}

\item{num}{a numeric value}
}
\value{
no return value
}
\description{
Multiply a matrix and a numeric value and allocate the result to the matrix memory.
}
