% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matXrow}
\alias{matXrow}
\title{Matrix X Row vector}
\usage{
matXrow(mat, vec)
}
\arguments{
\item{mat}{a numeric matrix}

\item{vec}{a numeric vector}
}
\value{
no return value
}
\description{
Multiply a matrix and a row vector and allocate the result to the matrix memory.
}
