% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Extract meta information about an object}
\usage{
meta(x)
}
\arguments{
\item{x}{An R object (e.g., data.frame).}
}
\value{
A meta object summarizing information (see \code{\link[=meta.data.frame]{meta.data.frame()}}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Generic function to display meta information about an object.
Currently supports data.frame objects with a specific method.
}
\examples{
\donttest{meta(mtcars)}

}
