% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{min_by_col_nm}
\alias{min_by_col_nm}
\title{max, min, sum by column names}
\usage{
min_by_col_nm(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric matrix}

\item{na.rm}{a boolean value removing NA or not}
}
\value{
A matrix of minimum values shrinked by column names
}
\description{
max, min, sum of values on each column name of a matrix.
}
\examples{
# max by col nm
\donttest{x <- matrix(c(1:9), ncol = 3)
set_dim_nm(x, list(c("a", "a", "b"), c("a", "b", "b")))
max_by_col_nm(x)}

# min by col nm
\donttest{x <- matrix(c(1:9), ncol = 3)
set_dim_nm(x, list(c("a", "a", "b"), c("a", "b", "b")))
min_by_col_nm(x)}

# sum by col nm
\donttest{x <- matrix(c(1:9), ncol = 3)
set_dim_nm(x, list(c("a", "a", "b"), c("a", "b", "b")))
sum_by_col_nm(x)}

}
