% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{mkdir}
\alias{mkdir}
\title{mkdir}
\usage{
mkdir(
  folders = c("dev", "data", "info", "inst", "output", "R", "raw", "report", "rules")
)
}
\arguments{
\item{folders}{A string vector specifying folder names}
}
\value{
no return, this function makes folders. if the folders already exist,
no effect
}
\description{
Same function of linux mkdir command
}
