% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{mostfreq}
\alias{mostfreq}
\title{Most frequent value (mode, modal value)}
\usage{
mostfreq(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{A boolean value removing na or not}
}
\value{
the most frequent value vector and its frequency
}
\description{
Get the most frequent value.
}
\examples{
# get the most frequent values
\donttest{x <- c(1, 1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 5)
mostfreq(x)}

}
