% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_comma}
\alias{paste_comma}
\title{Paste vector elements with commas}
\usage{
paste_comma(x, newline = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{newline}{Logical; if \code{TRUE}, each element is placed on a new line.}
}
\value{
No return value, prints to the console.
}
\description{
Combine elements of a vector into a string with commas.
Optionally insert newlines between elements for readability.
}
\examples{
# paste elements with commas
\donttest{paste_comma(names(mtcars))}

}
