% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_list}
\alias{paste_list}
\title{Paste vectors of a list}
\usage{
paste_list(x, sep = "|")
}
\arguments{
\item{x}{a list with same length vectors or data frame column vectors you want to paste.}

\item{sep}{a character string to separate the terms.}
}
\value{
a vector pasted
}
\description{
Paste vectors of equal length in a list or data.frame
}
\examples{
# paste length and width of iris
iris$size <- paste_list(iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")])
head(iris)

}
