% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{paste_sort_uni_str}
\alias{paste_sort_uni_str}
\title{Paste sorted unique strings}
\usage{
paste_sort_uni_str(x, collapse = "|")
}
\arguments{
\item{x}{A character vector.}

\item{collapse}{A single string used to separate elements (not \code{NA_character_}).
Default: \code{"|"}.}
}
\value{
A length-1 character vector.
}
\description{
Collapse \strong{sorted} unique, non-missing elements into a single string.
}
\examples{
# paste sorted unique string
\donttest{paste_sort_uni_str(c("a", "a", "c", "b", "b", "d", "e"))}

}
