% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{paste_uni_str}
\alias{paste_uni_str}
\title{Paste unique strings}
\usage{
paste_uni_str(x, collapse = "|")
}
\arguments{
\item{x}{a string vector}

\item{collapse}{an optional character string to separate the results. Not
\code{\link{NA_character_}}. When collapse is a string, the result is always a string
(\code{\link{character}} of length 1). default "|"}
}
\value{
a string vector
}
\description{
Paste unique strings
}
\examples{
# paste unique string
\donttest{paste_uni_str(c("a", "a", "c", "b", "b", "d", "e"))}

}
