% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{pull_str}
\alias{pull_str}
\title{Pull a first string having a specific pattern.}
\usage{
pull_str(pattern, x, ignore.case = TRUE)
}
\arguments{
\item{pattern}{a string containing a \link{regular expression}}

\item{x}{a string vector}

\item{ignore.case}{if \code{FALSE}, the pattern matching is case sensitive and if \code{TRUE}, case is ignored during matching.}
}
\value{
a string vector
}
\description{
Pull a first string having a specific pattern.
}
\examples{
# pull a first string having a specific pattern
\donttest{pull_str(pattern = "c", c("a|b|c", "a|c|c"))}

}
