% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{pull_str_all}
\alias{pull_str_all}
\title{Pull all the strings having a specific pattern.}
\usage{
pull_str_all(pattern, x, collapse = "|", ignore.case = TRUE)
}
\arguments{
\item{pattern}{a string containing a \link{regular expression}}

\item{x}{a string vector}

\item{collapse}{an optional character string to separate the results. Not
\code{\link{NA_character_}}. When collapse is a string, the result is always a string
(\code{\link{character}} of length 1). default "|"}

\item{ignore.case}{if \code{FALSE}, the pattern matching is case sensitive and if \code{TRUE}, case is ignored during matching.}
}
\value{
a string vector
}
\description{
Pull all the strings having a specific pattern.
}
\examples{
# pull all the strings having a specific pattern
\donttest{pull_str_all(pattern = "c", c("a|b|c", "a|c|c"))}

}
