% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quote_comma}
\alias{quote_comma}
\title{Quote and paste symbols with commas}
\usage{
quote_comma(..., newline = FALSE)
}
\arguments{
\item{...}{Unquoted variable names or expressions.}

\item{newline}{Logical; if \code{TRUE}, print each element on a new line.}
}
\value{
No return value, prints to the console.
}
\description{
Capture unquoted symbols, wrap them in quotes, and print them
as a comma-separated character vector.
}
\examples{
\donttest{quote_comma(mpg, cyl, disp, hp, drat)} # c("mpg", "cyl", "disp", "hp", "drat")

}
