% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_rds}
\alias{read_rds}
\title{read_rds}
\usage{
read_rds(file, refhook = NULL)
}
\arguments{
\item{file}{a \link{connection} or the name of the file where the R object is saved
to or read from.}

\item{refhook}{a hook function for handling reference objects.}
}
\value{
an \R object
}
\description{
read_rds is almost same as readRDS except for a pointer.
}
\examples{
# compare pointer values
\dontrun{data <- copy(women)
data.table::setDT(data)
saveRDS(data, "data.rds")
df <- readRDS("data.rds")
dt <- read_rds("data.rds")
attributes(df)$.internal.selfref # <pointer: (nil)>
attributes(dt)$.internal.selfref}

}
