% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{regex_attr}
\alias{regex_attr}
\title{Matched attributes using regular expressions}
\usage{
regex_attr(x, name)
}
\arguments{
\item{x}{any objects; e.g, list, columns of a data.frame or data.table}

\item{name}{the regular expression attribute names.}
}
\value{
matched attributes' names
}
\description{
Find matched attributes using regular expressions.
}
\examples{
# matched attributes using regular expressions
\donttest{regex_attr(x, "class|names")}

}
