% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_cols}
\alias{regex_cols}
\title{Find columns by regular expression}
\usage{
regex_cols(df, pattern)
}
\arguments{
\item{df}{A data.frame.}

\item{pattern}{A character string containing a regular expression.}
}
\value{
A character vector of matching column names.
}
\description{
Return column names that match a regular expression.
}
\examples{
\donttest{
df <- data.frame(col_a = c(1, 2, 3), col_b = c("A", "B", "C"), col_c = c(4, 5, 6))
regex_cols(df, pattern = c("a|c"))
}

}
