% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{rep_col}
\alias{rep_col}
\title{Replicate columns of a matrix}
\usage{
rep_col(x, each)
}
\arguments{
\item{x}{A matrix.}

\item{each}{Integer; number of times to repeat each column.}
}
\value{
A matrix with \code{ncol(x) * each} columns.
}
\description{
Repeat each column of a matrix \code{each} times, returning a matrix with
replicated columns (column-wise Kronecker-style replication).
}
\examples{
\donttest{
m <- matrix(1:6, nrow = 2)
rep_col(m, each = 3)
}

}
