% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_a_with_b}
\alias{replace_a_with_b}
\title{Replace A with B}
\usage{
replace_a_with_b(df, cols, a, b)
}
\arguments{
\item{df}{a data frame}

\item{cols}{a string vector specifying columns}

\item{a}{a string}

\item{b}{a string}
}
\value{
no return value
}
\description{
Replace string A with string B in a memory-efficient way
}
\examples{
\dontrun{
df <- data.table(x = c("A", "A", "C"), y = c("A", "C", "C"))
set_ptr(df)
data.table::address(df)
replace_a_with_b(df, a = "A", b = "B")
data.table::address(df)
df}

}
