% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_empty_with_na}
\alias{replace_empty_with_na}
\title{Replace empty with NA}
\usage{
replace_empty_with_na(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
no return value
}
\description{
Replace empty string like "" with NA_character_ in a memory-efficient way
}
\examples{
\donttest{df <- data.frame(x = c("A", "B", ""), y = c(1, NA, 3), z = c("", "E", ""))
set_ptr(df)
data.table::address(df)
replace_empty_with_na(df)
data.table::address(df)
df}

}
