% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_na_with_empty}
\alias{replace_na_with_empty}
\title{Replace NA with empty}
\usage{
replace_na_with_empty(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
no return value
}
\description{
Replace NA_character_ with "" in a memory-efficient way
}
\examples{
\donttest{df <- data.frame(x = c("A", "B", NA), y = c(1, NA, 3), z = c(NA, "E", NA))
pryr::address(df)
replace_na_with_empty(df)
pryr::address(df)
df}

}
