% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_na_with_zero}
\alias{replace_na_with_zero}
\title{Replace NA with zero}
\usage{
replace_na_with_zero(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
no return value
}
\description{
Replace NA_integer_ or NA_real_ values with zero in a memory-efficient way
}
\examples{
\donttest{df <- data.frame(x = c(1, NA, 3), y = c("A", "B", NA), z = c(NA, 5, NA))
pryr::address(df)
replace_na_with_zero(df)
pryr::address(df)
df}

}
