% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{replace_zen_to_han}
\alias{replace_zen_to_han}
\title{Convert full-width to half-width for data.table columns}
\usage{
replace_zen_to_han(df)
}
\arguments{
\item{df}{A data.table. Modified in place.}
}
\value{
The updated data.table.
}
\description{
Renames columns using \code{\link[=zen_to_han]{zen_to_han()}} and also converts the contents of any
columns that contain Japanese text (as determined by \code{\link[=is_japanese]{is_japanese()}}).
}
\examples{
\dontrun{
df <- data.table::data.table("Ａ" = c("Ａ", "Ｂ"))
replace_zen_to_han(df)
}

}
