% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{reverse}
\alias{reverse}
\title{Reverse the order of elements in a vector}
\usage{
reverse(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of the same type as \code{x}, with elements in reverse order.
}
\description{
Returns the input vector with its elements in reverse order.
}
\examples{
\donttest{
# Reverse a numeric vector
x <- c(1:10)
reverse(x)
}

}
