% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{rm_punct}
\alias{rm_punct}
\title{Remove punctuations}
\usage{
rm_punct(df, cols, pattern = "(?!\\\\*)[[:punct:]]")
}
\arguments{
\item{df}{a data.frame}

\item{cols}{a string vector specifying columns}

\item{pattern}{a string containing a \link{regular expression}}
}
\value{
no return values
}
\description{
Remove punctuations.
}
\examples{
# remove punctuations
\donttest{df <- data.frame(x = c("A3-", "$+_B", "C+_&"), y = c("123", "R&", "4q_++"))
set_ptr(df)
rm_punct(df)
df}

}
