% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{rowvec}
\alias{rowvec}
\alias{colvec}
\title{Convert a vector into a row or column vector}
\usage{
rowvec(x)

colvec(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A matrix with either one row (\code{rowvec()}) or one column (\code{colvec()}),
preserving names where available.
}
\description{
These functions reshape a one-dimensional vector into either a row vector
(1 × \emph{n}) or a column vector (\emph{n} × 1).
}
\examples{
\donttest{
# Convert into a row vector
rowvec(c(1, 2, 3, 4, 5))

# Convert into a column vector
colvec(c(1, 2, 3, 4, 5))
}

}
