% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{save_plot_xlsx}
\alias{save_plot_xlsx}
\title{Save ggplot object(s) to an Excel workbook}
\usage{
save_plot_xlsx(
  plot,
  file,
  rc = c(1L, 1L),
  width = 12,
  height = 6,
  dpi = 300,
  overwrite = FALSE
)
}
\arguments{
\item{plot}{A \code{ggplot} object or a named/unnamed list of \code{ggplot} objects.}

\item{file}{A file path to the target Excel workbook.}

\item{rc}{A length-2 integer vector giving the starting row and column
(default is \code{c(1L, 1L)}).}

\item{width}{Image width (in inches).}

\item{height}{Image height (in inches).}

\item{dpi}{Resolution of the image (dots per inch).}

\item{overwrite}{Logical; if TRUE, overwrite an existing workbook.}
}
\value{
No return value, called for side effects (saves an Excel file).
}
\description{
Save one or more \code{ggplot} objects to an Excel workbook. A single \code{ggplot}
object can be supplied directly, or multiple plots can be passed in a list.
Each plot is written to a separate worksheet.
}
\examples{
\dontrun{
library(ggplot2)

p1 <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p2 <- ggplot(mtcars, aes(hp, qsec)) + geom_point()

# Single plot
save_plot_xlsx(p1, "single.xlsx")

# Multiple plots
save_plot_xlsx(list(Scatter1 = p1, Scatter2 = p2), "multi.xlsx")
}

}
