% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{sec_to_hms}
\alias{sec_to_hms}
\title{Convert seconds to HH:MM:SS.subsec string}
\usage{
sec_to_hms(sec)
}
\arguments{
\item{sec}{Numeric seconds.}
}
\value{
A character string in the form \code{HH:MM:SS.subsec}.
}
\description{
Utility to format a numeric number of seconds as a clock-style string with
subseconds preserved.
}
\keyword{internal}
