% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_attr}
\alias{set_attr}
\title{Set attributes}
\usage{
set_attr(x, name, value)
}
\arguments{
\item{x}{any objects; e.g, list, columns of a data.frame or data.table}

\item{name}{the character attribute name.}

\item{value}{the value to assign to the attribute or \code{NULL} removes teh attribute, if present.}
}
\value{
the changed object (invisibly) for use in compound statements.
}
\description{
setattr functions re-exported from \code{data.table}.
}
\examples{
# set attributes
\dontrun{df <- data.frame(a = 1:3, b = 4:6)
set_attr(df, "flag", TRUE)
attr(df, "flag")}

}
