% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{set_band}
\alias{set_band}
\title{Deprecated: set_band()}
\usage{
set_band(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=add_band]{add_band()}}.}
}
\value{
Same return value as \code{\link[=add_band]{add_band()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=add_band]{add_band()}} instead.
}
\seealso{
\code{\link[=add_band]{add_band()}}
}
