% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_col_lower}
\alias{set_col_lower}
\alias{set_col_upper}
\title{Change columns from uppercase to lowercase or from lowercase to uppercase}
\usage{
set_col_lower(x)

set_col_upper(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
no return values
}
\description{
Change columns from uppercase to lowercase or from lowercase to uppercase
}
\examples{
# Change columns case
\donttest{df <- mtcars
set_col_upper(df)
set_col_lower(df)}

}
