% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{set_dim_nm}
\alias{set_dim_nm}
\alias{set_row_nm}
\alias{set_col_nm}
\title{Set dimension, row, column names}
\usage{
set_dim_nm(x, dimnames)

set_row_nm(x, rownames)

set_col_nm(x, colnames)
}
\arguments{
\item{x}{A matrix}

\item{dimnames, rownames, colnames}{A list of dimension names\cr
A vector of row names\cr
A vector of column names}
}
\value{
no return value
}
\description{
Set dimension, row, column names to a matrix.
}
\examples{
# set dimension names
\donttest{x <- matrix(c(1:9), ncol = 3)
set_dim_nm(x, list(c("a", "a", "b"), c("a", "b", "b")))}

# set row names
\donttest{x <- matrix(c(1:9), ncol = 3)
set_row_nm(x, c("a", "a", "b"))}

# set column names
\donttest{x <- matrix(c(1:9), ncol = 3)
set_col_nm(x, c("a", "b", "b"))}

}
