% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dt}
\alias{set_dt}
\title{Set data.frame to data.table}
\usage{
set_dt(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
No return value.
}
\description{
Set data.frame to data.table class.
}
\examples{
# set data.frame to data.table
\donttest{set_dt(iris)}

}
