% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_labels}
\alias{set_labels}
\title{Set labels}
\usage{
set_labels(df, labels, cols)
}
\arguments{
\item{df}{a data.frame}

\item{labels}{a string vector specifying labels to describe columns}

\item{cols}{a string vector specifying columns}
}
\description{
Set column labels for a data frame.
}
\examples{
# set labels
\dontrun{df <- data.frame(Q1 = c(0, 1, 1), Q2 = c(1, 0, 1))
set_labels(df, labels = c("Rainy?", "Umbrella?"))
View(df)}

}
