% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ptr}
\alias{set_ptr}
\title{Set external pointer}
\usage{
set_ptr(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
No return value.
}
\description{
Set external pointer
}
\examples{
# set pointer
\donttest{set_ptr(iris)}

}
