% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{set_stat_by}
\alias{set_stat_by}
\title{Set statistic by groups}
\usage{
set_stat_by(df, group_var, value_var, fun = cumsum, prefix = "c")
}
\arguments{
\item{df}{a data.frame}

\item{group_var}{names of columns for grouping}

\item{value_var}{names of columns to be applied for statistics.}

\item{fun}{a function applying to columns}

\item{prefix}{a new column's perfix string (if the value_var is "loss", the new column is "closs")}
}
\value{
no return values.
}
\description{
Set statistic column by groups (mainly cumulative function like \code{\link[=cumsum]{cumsum()}}, \code{\link[=cumprod]{cumprod()}}, \code{\link[=cummax]{cummax()}}, \code{\link[=cummin]{cummin()}})
}
\examples{
# set statistics by groups
\dontrun{
set_ptr(mtcars)
set_stat_by(mtcars, .(cyl, vs), value_var = hp, fun = cumsum)
mtcars[, c("cyl", "vs", "chp")]}

}
