% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_tibble}
\alias{set_tibble}
\title{Set data frame to tibble}
\usage{
set_tibble(x, envir = globalenv())
}
\arguments{
\item{x}{data.frame}

\item{envir}{the \link{environment} to use. See ‘Details’.}
}
\value{
No return value.
}
\description{
Set data frame to tibble class.
}
\examples{
# set data.frame to tibble
\donttest{set_tibble(iris)}

}
