% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeof.R
\name{sizeof}
\alias{sizeof}
\title{Object size}
\usage{
sizeof(x, unit)
}
\arguments{
\item{x}{object vector, data.frame, environment and etc}

\item{unit}{size unit \code{B}, \code{KB}, \code{MB}, \code{GB} default \code{KB}}
}
\description{
This function calculates the object size
}
\examples{
\dontrun{
sizeof(x = sys.frame(), unit = "kb")
}
}
