% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{split_str}
\alias{split_str}
\title{Split strings}
\usage{
split_str(x, split = "|")
}
\arguments{
\item{x}{A character vector.}

\item{split}{A single string delimiter.}
}
\value{
A list of character vectors (one element per input string).
}
\description{
Split each input string on a delimiter.
}
\examples{
# split strings
\donttest{split_str(c("a|b|a", "c|d|c"), split = "|")}

}
