% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timeit}
\alias{timeit}
\title{Time it}
\usage{
timeit(expr)
}
\arguments{
\item{expr}{\R code expression}
}
\description{
Time \R code expression
}
\examples{
# time it
\donttest{timeit(for (i in 1:1000000) i)}

}
