% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{to_a1_col}
\alias{to_a1_col}
\title{Convert column numbers to Excel letters}
\usage{
to_a1_col(x)
}
\arguments{
\item{x}{Numeric column number.}
}
\value{
String with Excel column letters.
}
\description{
Converts column numbers to Excel column letters (1=A, 2=B, 27=AA, etc.).
}
\examples{
\donttest{
# Convert column numbers to letters
to_a1_col(1)   # Returns "A" (1st column)
to_a1_col(3)   # Returns "C" (3rd column)
to_a1_col(26)  # Returns "Z" (26th column)
to_a1_col(27)  # Returns "AA" (27th column)
}

}
