% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{to_r1c1_col}
\alias{to_r1c1_col}
\title{To a R1C1 format column}
\usage{
to_r1c1_col(x)
}
\arguments{
\item{x}{A string specifying a cell or a column}
}
\value{
A numeric
}
\description{
Convert a cell or column string from A1 to R1C1 format.
}
\examples{
# convert a cell or column string from A1 to R1C1 format
\donttest{
to_r1c1_col("C33")
to_r1c1_col("ABC")
to_r1c1_col("ABC123")}

}
