% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{trim_ws}
\alias{trim_ws}
\title{Trim white space}
\usage{
trim_ws(df, ws = "[ \\t\\r\\n]")
}
\arguments{
\item{df}{a data frame}

\item{ws}{a white space \link{regular expression}}
}
\value{
no return value
}
\description{
Trim white space
}
\examples{
\donttest{df <- data.frame(x = c(" A", "B ", " C "), y = c(1, 2, 3))
pryr::address(df)
trim_ws(df)
pryr::address(df)
df}

}
