% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{type.data.frame}
\alias{type.data.frame}
\title{Class and type information for data frames}
\usage{
\method{type}{data.frame}(x)
}
\arguments{
\item{x}{A data.frame.}
}
\value{
A data.table with columns:
\itemize{
\item \code{column}: column name
\item \code{class}: column class
\item \code{type}: storage type
}
}
\description{
Summarize class and storage type of each column in a data frame.
}
