% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{type}
\alias{type}
\title{Class and type information}
\usage{
type(x)
}
\arguments{
\item{x}{object vector, data.frame, environment and etc}
}
\value{
class and type information data frame
}
\description{
Show class and type information.
}
\examples{
\donttest{type(cars)}

}
