% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{unilen}
\alias{unilen}
\title{length of a unique vector}
\usage{
unilen(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A length of a unique vector
}
\description{
Calculate length of a unique vector.
}
\examples{
# length of unique vector
\donttest{x <- c(1, 1, 2, 3, 4, 5, 5)
reverse(x)}

}
