% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_cols}
\alias{valid_cols}
\title{Validate columns}
\usage{
valid_cols(df, cols)
}
\arguments{
\item{df}{a data.frame}

\item{cols}{a string vector specifying columns}
}
\value{
no return value
}
\description{
Does the data frame contain all columns?
}
\examples{
# different columns
\donttest{valid_cols(mtcars, c("mpg", "cyl", "disp", "hp", "drat"))}

}
