% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{view_data_image}
\alias{view_data_image}
\alias{save_data_image}
\title{View a html data image}
\usage{
view_data_image(
  data,
  caption = "Table.1",
  footnote = NULL,
  digits = 2,
  full_width = FALSE,
  html_font = "Comic Sans MS"
)

save_data_image(
  data,
  caption = "Table.1",
  footnote = NULL,
  digits = 2,
  full_width = FALSE,
  html_font = "Comic Sans MS",
  zoom = 1.5,
  file
)
}
\arguments{
\item{data}{a data.frame}

\item{caption}{The table caption.}

\item{footnote}{A vector of footnote texts, Footnotes here will be labeled with special symbols.
The vector here should not have more than 20 elements.}

\item{digits}{Maximum number of digits for numeric columns, passed to round(). This can also be a vector of length ncol(x), to set the
number of digits for individual columns.}

\item{full_width}{A TRUE or FALSE variable controlling whether the HTML table should have the preferable format for
\code{full_width}. If not specified, a HTML table will have full width by default but this option will be set to \code{FALSE} for a LaTeX table}

\item{html_font}{A string for HTML css font.}

\item{zoom}{A number specifying the zoom factor. A zoom factor of 2 will result in twice as many pixels vertically and horizontally. Note that
using 2 is not exactly the same as taking a screenshot on a HiDPI (Retina) device: it is like increasing the zoom to 200 doubling the
height and width of the browser window. This differs from using a HiDPI device because some web pages load different, higher-
resolution images when they know they will be displayed on a HiDPI device (but using zoom will not report that there is a HiDPI
device).}

\item{file}{A file path}
}
\value{
no return value.
}
\description{
View a html data image (Viewer in RStudio).
}
\examples{
# plot a data image
\dontrun{
plot_data_image(head(data), height = .5)}

}
