% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{view_table_html}
\alias{view_table_html}
\title{View a table as an HTML table}
\usage{
view_table_html(
  table,
  caption = "Table.1",
  footnote = NULL,
  digits = 2,
  full_width = FALSE,
  html_font = "Comic Sans MS"
)
}
\arguments{
\item{table}{A data.frame, matrix, or table object to be rendered as an HTML table.}

\item{caption}{The table caption.}

\item{footnote}{A vector of footnote texts, Footnotes here will be labeled with special symbols.
The vector here should not have more than 20 elements.}

\item{digits}{Maximum number of digits for numeric columns, passed to round(). This can also be a vector of length ncol(x), to set the
number of digits for individual columns.}

\item{full_width}{A TRUE or FALSE variable controlling whether the HTML table should have the preferable format for
\code{full_width}. If not specified, a HTML table will have full width by default but this option will be set to \code{FALSE} for a LaTeX table}

\item{html_font}{A string for HTML css font.}
}
\value{
No return value.
}
\description{
Render a table into an HTML table and display it in the RStudio Viewer
(or browser if available).
}
\examples{
\dontrun{
# 2x2 Table with totals
set.seed(1)
smoker  <- sample(c("Yes","No"),  100, TRUE)
disease <- sample(c("Yes","No"),  100, TRUE)
tab <- table(Smoker = smoker, Disease = disease)
tab_tot <- addmargins(tab, margin = c(1, 2))
view_table_html(tab_tot, caption = "2x2 Table with Totals")
}

}
