% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write data in an excel file}
\usage{
write_xlsx(data, file, rc = c(1L, 1L), overwrite = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{file}{A file path}

\item{rc}{A vector of starting point row and column, default c(1L, 1L)}

\item{overwrite}{A boolean value that overwrite or not}
}
\value{
no return
}
\description{
Write data in an excel file.
}
\examples{
# write xlsx file
\dontrun{
write_xlsx(list(data1, data2), "data.xlsx")}

}
