% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{yearmon}
\alias{yearmon}
\title{Format dates as YYYYMM}
\usage{
yearmon(x)
}
\arguments{
\item{x}{A Date vector.}
}
\value{
A character vector in \code{YYYYMM} format (e.g., "202312").
}
\description{
Extract year and month from dates as a compact \code{YYYYMM} character string.
}
\examples{
\donttest{
# Get year-month
x <- as.Date(c("1999-12-31", "2000-01-01"))
yearmon(x)
}

}
