% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{zen2han4dat}
\alias{zen2han4dat}
\title{Zenkaku to Hankaku for a data frame}
\usage{
zen2han4dat(df)
}
\arguments{
\item{df}{a string vector}
}
\value{
a string vector
}
\description{
Covert Zenkaku columns to Hankaku columns
}
