% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{zen_to_han}
\alias{zen_to_han}
\title{Zenkaku to Hankaku}
\usage{
zen_to_han(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with half-width forms where applicable.
}
\description{
Converts full-width forms to half-width for characters where such a mapping
exists. Input is converted to UTF-8 if necessary.
}
\examples{
\dontrun{
zen_to_han(c("ＡＢＣ", "１２３", "ｱｲｳ", "テスト　"))
}

}
