% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{zeros}
\alias{zeros}
\title{Zeros}
\usage{
zeros(dim)
}
\arguments{
\item{dim}{a numeric vector specifying dimensions}
}
\value{
a matrix with zeros
}
\description{
Sample matrix containing only zeros
}
\examples{
# zeros
zeros(c(5, 5))

}
