% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{add_mon}
\alias{add_mon}
\title{Add months to a date}
\usage{
add_mon(date, mon)
}
\arguments{
\item{date}{A date-like object (coercible via \code{\link[=as.POSIXlt]{as.POSIXlt()}}).}

\item{mon}{Integer number of months to add (can be negative).}
}
\value{
A Date with the months added.
}
\description{
Add a specified number of months to a date-like object. Negative values
subtract months. Uses base rollover via POSIXlt.
}
\examples{
\donttest{
# Add months
date <- Sys.Date()
add_mon(date, 3)
}

}
