% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_comma}
\alias{as_comma}
\title{Format numbers with commas}
\usage{
as_comma(x)
}
\arguments{
\item{x}{A numeric or integer vector.}
}
\value{
A character vector of formatted numbers.
}
\description{
Convert a numeric or integer vector into a character vector
formatted with commas as thousands separators.
}
\examples{
# format numbers with commas
\donttest{as_comma(c(123456, 234567))}

}
