% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_class}
\alias{assert_class}
\title{Assert object class}
\usage{
assert_class(x, class)
}
\arguments{
\item{x}{An R object.}

\item{class}{A character vector of class names to check against.}
}
\value{
No return value. Called for side effects.
}
\description{
Check that an object inherits from the expected class.
Throws an error if the assertion fails.
}
\examples{
\donttest{assert_class(cars, "data.frame")}

}
