% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_class}
\alias{assert_class}
\title{assert class}
\usage{
assert_class(obj, class)
}
\arguments{
\item{obj}{object}

\item{class}{A class string}
}
\value{
no return
}
\description{
Check the class of object.
}
