% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_type}
\alias{assert_type}
\title{Assert type}
\usage{
assert_type(obj, type)
}
\arguments{
\item{obj}{object}

\item{type}{object type}
}
\value{
no return
}
\description{
Assert object type.
}
\examples{
# assert object type
\donttest{assert_type("x", "character")}

}
