% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{bmonth}
\alias{bmonth}
\alias{emonth}
\title{Beginning and end of the month}
\usage{
bmonth(date)

emonth(date)
}
\arguments{
\item{date}{A Date.}
}
\value{
A Date object representing the beginning (bmonth) or end (emonth)
of the month
}
\description{
Get the beginning date (first day) or end date (last day) of the month.
}
\examples{
\donttest{
# Beginning of month
bmonth(Sys.Date())

# End of month
emonth(Sys.Date())
}

}
