% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{combine_overlapping_date_range}
\alias{combine_overlapping_date_range}
\title{Combine overlapping date ranges}
\usage{
combine_overlapping_date_range(
  df,
  id_var,
  merge_var,
  from_var,
  to_var,
  interval = 0,
  collapse = "|"
)
}
\arguments{
\item{df}{a data.frame with data ranges}

\item{id_var}{id column names}

\item{merge_var}{column names to be collapsed}

\item{from_var}{a column name of start date}

\item{to_var}{a column name of end date}

\item{interval}{an interval of previous end date to next start date to be combined}

\item{collapse}{an optional character string to seperate the result of \code{merge_var}}
}
\value{
\code{data.frame} with no overlapping date ranges.
}
\description{
Combine multiple overlapping date ranges.
}
\examples{
# combine overlapping date ranges
id <- c("A", "A", "B")
work <- c("cleansing", "analysis", "cleansing")
sdate <- as.Date(c("2022-03-01", "2022-03-05", "2022-03-08"))
edate <- as.Date(c("2022-03-06", "2022-03-09", "2022-03-10"))
df <- data.frame(id = id, work = work, sdate = sdate, edate = edate)
combine_overlapping_date_range(df, id, work, sdate, edate, interval = 0)

}
