% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{combine_overlapping_date_range}
\alias{combine_overlapping_date_range}
\title{Deprecated: combine_overlapping_date_range}
\usage{
combine_overlapping_date_range(...)
}
\arguments{
\item{...}{Additional arguments passed to combine_overlapping_date_range().}
}
\value{
Same return value as \code{\link[=collapse_date_ranges]{collapse_date_ranges()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link[=collapse_date_ranges]{collapse_date_ranges()}} instead.
}
\seealso{
\code{\link[=collapse_date_ranges]{collapse_date_ranges()}}
}
