% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{count_pattern}
\alias{count_pattern}
\title{Count pattern matches}
\usage{
count_pattern(pattern, x, ignore.case = FALSE)
}
\arguments{
\item{pattern}{A string containing a regular expression.}

\item{x}{A character vector.}

\item{ignore.case}{Logical; if \code{TRUE}, case is ignored. Default: \code{FALSE}.}
}
\value{
An integer vector of match counts (same length as \code{x}).
}
\description{
Count regex matches in each string.
}
\examples{
\donttest{count_pattern(pattern = "c", c("a|b|c", "a|c|c"))}

}
