% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{del_pattern}
\alias{del_pattern}
\title{Delete patterns}
\usage{
del_pattern(pattern, x)
}
\arguments{
\item{pattern}{A string containing a regular expression.}

\item{x}{A character vector.}
}
\value{
A character vector with matches removed.
}
\description{
Remove all regex matches from each string.
}
\examples{
\donttest{del_pattern(pattern = "c", c("abc", "acc"))}

}
