% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{devars}
\alias{devars}
\title{devars}
\usage{
devars(x)
}
\arguments{
\item{x}{a string, vector or list expression that can be a string vector}
}
\value{
a string vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function operates like \code{deparse(substitute(x))} inside the functions.
}
\examples{
# deparse(substitute(x))
\dontrun{devars(expression)
devars(c(expression, string))
devars(.(expression, string))
devars(list(expression, string))}
}
\keyword{internal}
