% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_cols}
\alias{diff_cols}
\title{Columns not in a set}
\usage{
diff_cols(df, cols)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{A character vector of column names.}
}
\value{
A character vector of column names in \code{df} but not in \code{cols}.
}
\description{
Return the columns of a data frame that are \strong{not} in a specified list.
}
\examples{
\donttest{diff_cols(mtcars, c("mpg", "cyl", "disp", "hp", "drat"))}

}
