% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{dolock}
\alias{dolock}
\alias{unlock}
\title{Lock and unlock data objects}
\usage{
dolock(obj, key)

unlock(obj, key)
}
\arguments{
\item{obj}{An R object to be locked or unlocked.}

\item{key}{A character string specifying the password used for locking/unlocking.}
}
\value{
The locked (encrypted) object for \code{dolock()}, or the unlocked
(decrypted) object for \code{unlock()}.
}
\description{
\code{dolock()} encrypts and locks an object using a password key.
\code{unlock()} decrypts and restores a locked object with the same key.
}
\examples{
\dontrun{
## Lock an object
locked <- dolock("secret", "password")

## Unlock it
original <- unlock(locked, "password")
}

}
