% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{draw_line}
\alias{draw_line}
\title{Draw a horizontal line in the console}
\usage{
draw_line(width, mark = "=")
}
\arguments{
\item{width}{A numeric vector specifying the desired line width(s).
Defaults to the current console width (\code{getOption("width")}).}

\item{mark}{A single string giving the character(s) to repeat.
Defaults to \code{"="}.}
}
\value{
A character vector, each element being a line of repeated marks.
}
\description{
Create a line made of repeated characters, typically used as a visual
separator in console output. The line width defaults to the console
option \code{getOption("width")}.
}
\examples{
\donttest{
# Default line
cat(draw_line())

# Custom width
cat(draw_line(20))

# Custom mark
cat(draw_line(30, mark = "="))
}

}
\seealso{
\code{\link[cli:cat_line]{cli::cat_rule()}}
}
