% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{get_copied_dt}
\alias{get_copied_dt}
\title{Deprecated: get_copied_dt()}
\usage{
get_copied_dt(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.table copied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Convert a data.frame to data.table
}
\examples{
\donttest{
df <- data.frame(x = 1:3, y = c("a", "b", "c"))
get_copied_dt(df)
}

}
