% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_attr}
\alias{has_attr}
\title{Has attributes}
\usage{
has_attr(df, attr, error_raise = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{attr}{attribute names}

\item{error_raise}{a boolean whether to raise an error or not}
}
\value{
a boolean value
}
\description{
Whether the data has specific attributes
}
\examples{
# has attributes
\donttest{has_attr(mtcars, c("names", "class"))}

# raise an error
\dontrun{
has_attr(mtcars, c("names", "types"), error_raise = TRUE)}

}
