% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_ptr}
\alias{has_ptr}
\title{Has not a null pointer?}
\usage{
has_ptr(x, error_raise = TRUE)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a logical value whether to have not a null external pointer or not
}
\description{
Has not a null external pointer?
}
\examples{
# Has not null external pointer?
\donttest{has_ptr(iris)}

}
