% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{has_ptr}
\alias{has_ptr}
\title{Has not a null pointer?}
\usage{
has_ptr(df, error_raise = TRUE)
}
\arguments{
\item{df}{a data.frame}

\item{error_raise}{a logcial whether to raise an error or not}
}
\value{
a logical value whether to have not a null external pointer or not
}
\description{
Has not a null external pointer?
}
\examples{
# Has not null external pointer?
\donttest{has_ptr(iris, error_raise = FALSE)}

}
