% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_rows}
\alias{has_rows}
\title{Has rows}
\usage{
has_rows(df, error_raise = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{error_raise}{a logcial whether to raise an error or not}
}
\value{
a boolean value
}
\description{
Whether the data has rows
}
\examples{
# has rows
\dontrun{
df <- data.frame()
has_rows(df)}

# raise an error
\dontrun{
df <- data.frame()
has_rows(df, error_raise = TRUE)}

}
