% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icol}
\alias{icol}
\title{order numbers of columns}
\usage{
icol(x, cols)
}
\arguments{
\item{x}{a data.frame}

\item{cols}{string names of columns}
}
\value{
order numbers of columns
}
\description{
Get order numbers of columns
}
\examples{
# order numbers of columns
icol(mtcars, c("disp", "drat", "qsec", "am", "carb"))

}
