% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{image_xlsx}
\alias{image_xlsx}
\title{Deprecated: image_xlsx()}
\usage{
image_xlsx(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=save_image_xlsx]{save_image_xlsx()}}.}
}
\value{
Same return value as \code{\link[=save_image_xlsx]{save_image_xlsx()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link[=save_image_xlsx]{save_image_xlsx()}} instead.
}
\seealso{
\code{\link[=save_image_xlsx]{save_image_xlsx()}}
}
