% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_cols}
\alias{match_cols}
\title{Match columns}
\usage{
match_cols(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A character vector of column names.}
}
\value{
A character vector of matched column names (non-matching entries are dropped).
}
\description{
Return column names from a data frame that match a specified set.
}
\examples{
\donttest{
df <- data.frame(x = c(1, 2, 3), y = c("A", "B", "C"), z = c(4, 5, 6))
match_cols(df, c("x", "z"))
}

}
