% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mostfreq}
\alias{mostfreq}
\title{Most frequent value (mode, modal value)}
\usage{
mostfreq(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical. Should missing values (\code{NA}) be removed? Defaults to FALSE.}
}
\value{
The most frequent value vector and its frequency
}
\description{
Returns the most frequently occurring value (the statistical mode) in a vector.
}
\examples{
\donttest{
# Get the most frequent values
x <- c(1, 1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 5)
mostfreq(x)
}

}
