% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{paste_list}
\alias{paste_list}
\title{Paste vectors of a list}
\usage{
paste_list(x, sep = "|", na.rm = FALSE)
}
\arguments{
\item{x}{A list with same length vectors or data frame column vectors you want to paste.}

\item{sep}{A character string to separate the terms.}

\item{na.rm}{A logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
a vector pasted
}
\description{
Paste vectors of equal length in a list or data.frame
}
\examples{
\donttest{
iris$size <- paste_list(iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")])
head(iris)
}

}
