% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{plot_xlsx}
\alias{plot_xlsx}
\title{Deprecated: plot_xlsx()}
\usage{
plot_xlsx(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=save_plot_xlsx]{save_plot_xlsx()}}.}
}
\value{
Same return value as \code{\link[=save_plot_xlsx]{save_plot_xlsx()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link[=save_plot_xlsx]{save_plot_xlsx()}} instead.
}
\seealso{
\code{\link[=save_plot_xlsx]{save_plot_xlsx()}}
}
