% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_attr}
\alias{regex_attr}
\title{Find attributes by regular expression}
\usage{
regex_attr(x, name)
}
\arguments{
\item{x}{Any R object (e.g., \code{list}, \code{data.frame}, \code{data.table}).}

\item{name}{A character string containing a regular expression pattern
to match against attribute names.}
}
\value{
A character vector of matching attribute names.
}
\description{
Return the names of attributes in \code{x} whose names match a regular expression pattern.
}
\examples{
\donttest{
# Find attributes by regex
regex_attr(iris, "class|names")
}

}
