% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{rep_row}
\alias{rep_row}
\alias{rep_row.data.frame}
\alias{rep_row.data.table}
\alias{rep_row.matrix}
\title{Replicate rows of a matrix/data.frame/data.table}
\usage{
rep_row(x, ...)

\method{rep_row}{data.frame}(x, ...)

\method{rep_row}{data.table}(x, ...)

\method{rep_row}{matrix}(x, ...)
}
\arguments{
\item{x}{A matrix, data.frame or data.table.}

\item{...}{Arguments passed to \code{\link[base:rep]{base::rep()}} such as \code{times} and \code{each}.}
}
\value{
An object of the same general type as \code{x} with rows replicated.
}
\description{
Convenience S3 generic to replicate rows of tabular objects. For
each column, values are replicated according to arguments passed to
\code{\link[base:rep]{base::rep()}} (e.g., \code{times}, \code{each}).
}
\examples{
\donttest{
rep_row(iris, times = 3)
rep_row(iris, each  = 3)
}

}
