% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_a_with_b}
\alias{replace_a_with_b}
\title{Replace a specific value with another (in place)}
\usage{
replace_a_with_b(df, cols, a, b)
}
\arguments{
\item{df}{A data.table (modified in place).}

\item{cols}{A character vector of column names. Defaults to all character columns.}

\item{a}{A string to be replaced.}

\item{b}{Replacement string.}
}
\value{
The modified data.table.
}
\description{
Efficiently replace occurrences of string \code{a} with string \code{b} across selected
character columns in a data.table. If \code{cols} is missing, all character columns
are targeted.
}
\examples{
\dontrun{
# Replace A with B
df <- data.table(x = c("A", "A", "C"), y = c("A", "C", "C"))
data.table::address(df)
replace_a_with_b(df, a = "A", b = "B")
data.table::address(df)
}

}
