% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_empty_with_na}
\alias{replace_empty_with_na}
\title{Replace empty strings with NA (in place)}
\usage{
replace_empty_with_na(df, cols)
}
\arguments{
\item{df}{A data.table (modified in place).}

\item{cols}{A character vector of column names. Defaults to all character columns.}
}
\value{
The modified data.table.
}
\description{
Efficiently replace \code{""} with \code{NA_character_} in character columns of a
data.table. If \code{cols} is missing, all character columns are targeted.
}
\examples{
\dontrun{
# Replace "" with NA
df <- data.table(x = c("A", "B", ""), y = c(1, NA, 3), z = c("", "E", ""))
data.table::address(df)
replace_empty_with_na(df)
data.table::address(df)
}

}
