% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{reverse}
\alias{reverse}
\title{reverse a vector}
\usage{
reverse(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A reversed vector
}
\description{
Reverse a vector.
}
\examples{
# reverse a vector
\donttest{x <- c(1:10)
reverse(x)}

}
