% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{rm_cols}
\alias{rm_cols}
\title{Remove columns from a data.table}
\usage{
rm_cols(df, cols)
}
\arguments{
\item{df}{A data.table, modified in place.}

\item{cols}{Columns to remove, specified as a tidy-style
expression (e.g., \code{.(x, y)}, \code{list(x, y)}).}
}
\value{
The modified data.table, returned invisibly.
}
\description{
Remove one or more columns from a data.table by reference,
without making a copy. This is more memory-efficient than
creating a modified copy of the object.
}
\examples{
\dontrun{
# Remove columns
df <- data.table::as.data.table(mtcars)
address(df)
rm_cols(df, .(mpg, cyl, disp))
address(df)
}

}
