% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{rm_punct}
\alias{rm_punct}
\title{Remove punctuation (in place)}
\usage{
rm_punct(df, cols, pattern = "(?!\\\\*)[[:punct:]]")
}
\arguments{
\item{df}{A data.table (modified in place).}

\item{cols}{A character vector of column names. Defaults to all character columns.}

\item{pattern}{A regular expression describing punctuation to remove.}
}
\value{
The modified data.table.
}
\description{
Remove punctuation characters from character columns in a data.table. If
\code{cols} is missing, all character columns are targeted. The default pattern
avoids removing asterisks.
}
\examples{
\dontrun{
# Remove punctuations
df <- data.table(x = c("A3-", "$+_B", "C+_&"), y = c("123", "R&", "4q_++"))
data.table::address(df)
rm_punct(df)
data.table::address(df)
}

}
