% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{set_attr}
\alias{set_attr}
\title{Set attributes}
\usage{
set_attr(x, name, value)
}
\arguments{
\item{x}{Any R object (e.g., \code{list}, \code{data.frame}, \code{data.table})}

\item{name}{Character string specifying the attribute name}

\item{value}{Value to assign to the attribute, or \code{NULL} to remove the
attribute if it exists}
}
\value{
Returns the modified object invisibly for use in compound statements
}
\description{
A wrapper for \code{data.table::setattr()} to set object attributes.
}
\examples{
\donttest{
# Set attributes
df <- data.frame(a = 1:3, b = 4:6)
set_attr(df, "flag", TRUE)
attr(df, "flag")
}

}
