% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{set_attr_class}
\alias{set_attr_class}
\title{Set matched attributes' class}
\usage{
set_attr_class(x, name, class, regex = TRUE)
}
\arguments{
\item{x}{any objects; e.g, list, columns of a data.frame or data.table}

\item{name}{the attribute names.}

\item{class}{a string specifying class; e.g, c("data.table", "data.frame")}

\item{regex}{a logical specifying attributes finding method (a string vector or regular expression),
regex = \code{TRUE} is a default}
}
\value{
matched attributes' names
}
\description{
Set matched attributes' class.
}
\examples{
# set matched attributes' class
\donttest{df <- data.frame(x = 1:3, y = c("a", "b", "c"))
attr(df, "summary") <- data.frame(a = 4:6, b = c("d", "e", "f"))
attr(df, "summary")
set_attr_class(df, "sum", c("tbl_df", "tbl", "data.frame"), regex = TRUE)
attr(df, "summary")}

}
