% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_col_lower}
\alias{set_col_lower}
\alias{set_col_upper}
\title{Convert column names to lower or upper case}
\usage{
set_col_lower(df)

set_col_upper(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
The input data frame with column names modified in place.
Called for side effects.
}
\description{
Convenience functions to modify the case of all column names in a data frame.
These functions update the names \strong{in place} (using \code{data.table}).
}
\examples{
\donttest{
df <- mtcars

# Convert to upper case
set_col_upper(df)

# Convert to lower case
set_col_lower(df)

}

}
