% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dt}
\alias{set_dt}
\title{Set data.table function}
\usage{
set_dt(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
no return values.
}
\description{
setDT function re-exported from \code{data.table}.
}
\seealso{
\code{\link[=setDT]{setDT()}}
}
