% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{set_ptr}
\alias{set_ptr}
\title{Deprecated: set_ptr()}
\usage{
set_ptr(df, skip_shiny = TRUE)
}
\arguments{
\item{df}{A data.frame.}

\item{skip_shiny}{Logical; if \code{TRUE}, and a Shiny context is detected
(via \code{\link[=.is_shiny_running]{.is_shiny_running()}}), the function immediately returns
\code{data.table::setDT(df)}}
}
\value{
No return value. Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Set an external pointer.
}
\examples{
\donttest{
set_ptr(iris)
}

}
