% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{set_stat_by}
\alias{set_stat_by}
\title{Deprecated: set_stat_by()}
\usage{
set_stat_by(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=add_group_stats]{add_group_stats()}}.}
}
\value{
Same return value as \code{\link[=add_group_stats]{add_group_stats()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=add_group_stats]{add_group_stats()}} instead.
}
\seealso{
\code{\link[=add_group_stats]{add_group_stats()}}
}
