% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_tibble}
\alias{set_tibble}
\title{Set tibble function}
\usage{
set_tibble(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
no return values.
}
\description{
as_tibble function re-exported from \code{tibble}.
}
\seealso{
\code{\link[=as_tibble]{as_tibble()}}
}
