% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timeit}
\alias{timeit}
\title{Measure execution time of an expression}
\usage{
timeit(expr)
}
\arguments{
\item{expr}{An R expression to evaluate. Use \code{{ ... }} to time multiple lines.}
}
\description{
Evaluate an expression and return the elapsed time in \code{HH:MM:SS.subsec}
format. Intended for quick ad-hoc timing in scripts.
}
\examples{
# Time a simple loop
\donttest{timeit({ for (i in 1:1e6) i })}

}
