% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{traverse}
\alias{traverse}
\title{traverse two vectors}
\usage{
traverse(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A combined vector
}
\description{
Intersect the elements of two vectors.
}
\examples{
# traverse two vectors
\donttest{x <- y <- c(1:10)
traverse(x, y)}

}
