% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{type}
\alias{type}
\title{Inspect class and storage type}
\usage{
type(x)
}
\arguments{
\item{x}{An R object (e.g., data.frame).}
}
\value{
An R object summarizing class and storage type.
}
\description{
Generic function to show class and underlying storage type of objects.
Currently provides a method for data.frame.
}
\examples{
\donttest{type(cars)}

}
