% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{unilen}
\alias{unilen}
\title{Count unique elements}
\usage{
unilen(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An integer scalar: the count of distinct values in \code{x}.
}
\description{
Returns the number of distinct values in a vector. This is a lightweight,
fast alternative to \code{length(unique(x))}, implemented in C for performance.
}
\examples{
\donttest{
# Length of unique vector
x <- c(1, 1, 2, 3, 4, 5, 5)
unilen(x)
}

}
