% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_cols}
\alias{valid_cols}
\title{Validate that columns exist}
\usage{
valid_cols(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A character vector of column names to validate.}
}
\value{
No return value. Called for side effects.
}
\description{
Check whether a data frame contains all of the specified columns.
Throws an error if any are missing.
}
\examples{
\dontrun{valid_cols(mtcars, c("mpg", "cyl", "disp", "hp", "drat"))}

}
