% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{zeros}
\alias{zeros}
\title{Create a zero matrix}
\usage{
zeros(dim)
}
\arguments{
\item{dim}{Integer vector giving the dimensions of the result.}
}
\value{
A numeric matrix of zeros with dimensions specified by \code{dim}.
}
\description{
Generate a matrix filled with zeros of the specified dimensions.
}
\examples{
\donttest{
# 5x5 zero matrix
zeros(c(5, 5))
}

}
